% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetArchiveImage.R
\name{GetArchiveImage}
\alias{GetArchiveImage}
\title{Get an image from the archive}
\source{
\url{https://documentation.dataspace.copernicus.eu/APIs/SentinelHub/Process.html}
}
\usage{
GetArchiveImage(
  aoi,
  bbox,
  time_range,
  collection,
  script,
  format,
  mosaicking_order = c("mostRecent", "leastRecent", "leastCC"),
  file = NULL,
  pixels,
  resolution,
  buffer = 0,
  mask = FALSE,
  client,
  token,
  url = "https://services.sentinel-hub.com/api/v1/process"
)
}
\arguments{
\item{aoi}{sf or sfc object, typically a (multi)polygon, describing the Area of Interest}

\item{bbox}{numeric vector of four elements describing the bounding box of interest.

Only one of either aoi or bbox may be specified.}

\item{time_range}{scalar or vector (Date or character that can be converted to date) defining the time interval}

\item{collection}{character indicating which collection to search.
Must be one of the collections returned by \code{GetCollections}.}

\item{script}{character with tha evaluation script or the name of the file containing the script}

\item{format}{output format (PNG, JPG, TIFF)}

\item{mosaicking_order}{PARAM_DESCRIPTION. Default: c("mostRecent", "leastRecent", "leastCC")}

\item{file}{name of the file to save the image. If NULL the raster object is returened. Default: NULL}

\item{pixels}{PARAM_DESCRIPTION}

\item{resolution}{PARAM_DESCRIPTION}

\item{buffer}{width of the buffer (in meters) to retrieve image of enlarged area. Default: 0}

\item{mask}{logical indicating if the image should contain only pixels within aoi. Default: FALSE}

\item{client}{OAuth client object to use for authentication.}

\item{token}{OAuth token character string to use for authentication.}

\item{url}{character indicating the process endpoint. Default: Copernicus Data Space Ecosystem process endpoint}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[CDSE]{GetCollections}}
}
