% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStatistics.R
\name{GetStatistics}
\alias{GetStatistics}
\title{Get statistical values}
\source{
\url{https://documentation.dataspace.copernicus.eu/APIs/SentinelHub/Statistical.html}
}
\usage{
GetStatistics(
  aoi,
  bbox,
  time_range,
  collection,
  script,
  mosaicking_order = c("mostRecent", "leastRecent", "leastCC"),
  pixels,
  resolution,
  buffer = 0,
  percentiles = NULL,
  aggregation_period = 1L,
  lastIntervalBehavior = c("SKIP", "SHORTEN", "EXTEND"),
  as_data_frame = TRUE,
  client,
  token,
  url = getOption("CDSE.statistical_url")
)
}
\arguments{
\item{aoi}{sf or sfc object, typically a (multi)polygon, describing the Area of Interest.}

\item{bbox}{numeric vector of four elements describing the bounding box of interest.
    Specify with a coordinate pair on two (opposite) vertices of the bounding box rectangle.
    Coordinates need to be in longitude, latitude.

    Only one of either \code{aoi} or \code{bbox} may be specified.}

\item{time_range}{scalar or vector (Date or character that can be converted to date) defining the time interval.}

\item{collection}{character indicating which collection to search.
Must be one of the collections returned by \code{GetCollections}.}

\item{script}{a length one character string containing the evaluation script or the name of the file containing the script.}

\item{mosaicking_order}{character indicating the order in which tiles are overlapped from which the output result is mosaicked.
Must be one of "mostRecent", "leastRecent", or "leastCC". Default: "mostRecent"}

\item{pixels}{integer scalar or length-two vector indicating the request image width and height.
Values must be integers between 1 and 2500.}

\item{resolution}{numeric scalar or length-two vector indicating the spatial resolution of the request image
    in horizontal and vertical direction (in meters).

    Only one of the arguments "pixels" or "resolution" must be set at the same time.
    If the argument "pixels" or "resolution" is scalar, the same value is used for horizontal and vertical direction (width and height).}

\item{buffer}{numeric, width of the buffer to retrieve the image of enlarged area. Default: 0}

\item{percentiles}{numeric vector indicating which percentile values should be computed.
Default: NULL, don't compute any percentiles.}

\item{aggregation_period}{the length of the aggregation period in days. Default: 1}

\item{lastIntervalBehavior}{= c("SKIP", "SHORTEN", "EXTEND")}

\item{as_data_frame}{logical indicating if the result should be returned as data frame. Default: TRUE}

\item{client}{OAuth client object to use for authentication.}

\item{token}{OAuth token character string to use for authentication.

    Exactly one of either \code{client} or \code{token} must be specified. It is recommended to use \code{client}.}

\item{url}{character indicating the process endpoint. Default: Copernicus Data Space Ecosystem process endpoint}
}
\value{
\code{data.frame} or \code{list} with statistical values.
}
\description{
Retrieves the simple statistics for the area of interest calculated based on satellite imagery
    without having to download images.
}
\details{
The values are aggregated over the period (number of days) given by the \code{aggregation_period} argument.
   The default value 1 provides daily statistics. The statistics are returned only for the days when the data is
   available (the days of the satellite overpass).

   The scripts used for the Statistical API have some additional requirements: the \code{evaluatePixel()}
       function must, in addition to other output, always also return dataMask output.
       This output defines which pixels are excluded from calculations.
   For more information please visit the online documentation
   \url{https://documentation.dataspace.copernicus.eu/APIs/SentinelHub/Statistical.html}.

   If percentiles requested are 25, 50, and 75, the columns are renamed 'q1', 'median', and 'q3'.
}
\examples{
\dontrun{
dsn <- system.file("extdata", "centralpark.geojson", package = "CDSE")
aoi <- sf::read_sf(dsn, as_tibble = FALSE)
script_file <- system.file("scripts", "NDVI_CLOUDS_STAT.js", package = "CDSE")
daily_stats <- GetStatistics(aoi = aoi, time_range = c("2023-07-01", "2023-07-31"),
  collection = "sentinel-2-l2a", script = script_file, mosaicking_order = "leastCC",
  resolution = 100, aggregation_period = 1, client = OAuthClient)
weekly_stats <- GetStatistics(aoi = aoi, time_range = c("2023-07-01", "2023-07-31"),
  collection = "sentinel-2-l2a", script = script_file,mosaicking_order = "leastCC",
  resolution = 100, aggregation_period = 7, client = OAuthClient)
weekly_stats_extended <- GetStatistics(aoi = aoi, time_range = c("2023-07-01", "2023-07-31"),
  collection = "sentinel-2-l2a", script = script_file, mosaicking_order = "leastCC",
  resolution = 100, aggregation_period = 7, lastIntervalBehavior = "EXTEND", client = OAuthClient)
}
}
\seealso{
\code{\link[CDSE]{GetCollections}}, \code{\link[CDSE]{SearchCatalog}}
}
