% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueCatalog.R
\name{UniqueCatalog}
\alias{UniqueCatalog}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
UniqueCatalog(
  imageCatalog,
  by = c("areaCoverage", "tileCloudCover"),
  keep = names(imageCatalog)
)
}
\arguments{
\item{imageCatalog}{PARAM_DESCRIPTION}

\item{by}{PARAM_DESCRIPTION, Default: c("areaCoverage", "tileCloudCover")}

\item{keep}{PARAM_DESCRIPTION, Default: names(imageCatalog)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{aggregate}}
}
