% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetArchiveImage.R
\name{GetArchiveImage}
\alias{GetArchiveImage}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetArchiveImage(
  aoi,
  bbox,
  time_range,
  type,
  script,
  format,
  mosaicking_order = c("mostRecent", "leastRecent", "leastCC"),
  file = NULL,
  pixels,
  resolution,
  buffer = 0,
  mask = FALSE,
  client,
  token,
  url = "https://services.sentinel-hub.com/api/v1/process"
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{bbox}{PARAM_DESCRIPTION}

\item{time_range}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{script}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION}

\item{mosaicking_order}{PARAM_DESCRIPTION, Default: c("mostRecent", "leastRecent", "leastCC")}

\item{file}{PARAM_DESCRIPTION, Default: NULL}

\item{pixels}{PARAM_DESCRIPTION}

\item{resolution}{PARAM_DESCRIPTION}

\item{buffer}{PARAM_DESCRIPTION, Default: 0}

\item{mask}{PARAM_DESCRIPTION, Default: FALSE}

\item{client}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION}

\item{url}{PARAM_DESCRIPTION, Default: 'https://services.sentinel-hub.com/api/v1/process'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_transform}}, \code{\link[sf]{st_geometry}}, \code{\link[sf]{st_bbox}}, \code{\link[sf]{geos_unary}}, \code{\link[sf]{st_coordinates}}
 \code{\link[geojsonsf]{sfc_geojson}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr2]{request}}, \code{\link[httr2]{req_headers}}, \code{\link[httr2]{req_body}}, \code{\link[httr2]{req_auth_bearer_token}}, \code{\link[httr2]{req_oauth_client_credentials}}, \code{\link[httr2]{req_perform}}
 \code{\link[terra]{rast}}, \code{\link[terra]{crs}}, \code{\link[terra]{project}}, \code{\link[terra]{mask}}, \code{\link[terra]{writeRaster}}
}
