% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCollections.R
\name{GetCollections}
\alias{GetCollections}
\title{Get the list of available imagery collections}
\source{
\url{https://documentation.dataspace.copernicus.eu/APIs/SentinelHub/Catalog.html}
}
\usage{
GetCollections(url, as_data_frame = TRUE)
}
\arguments{
\item{url}{character, STAC collections endpoint}

\item{as_data_frame}{logical indicating if the result should be returned as data frame. Default: TRUE}
}
\value{
A \code{list} or a \code{data.frame} of all available imagery collections and their attributes.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[CDSE]{GetArchiveImage}}, \code{\link[CDSE]{SearchCatalog}}
}
