% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchCatalog.R
\name{SearchCatalog}
\alias{SearchCatalog}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SearchCatalog(
  aoi,
  bbox,
  from,
  to,
  collection,
  url,
  client,
  token,
  as_data_frame = TRUE,
  with_geometry = FALSE
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{bbox}{PARAM_DESCRIPTION}

\item{from}{PARAM_DESCRIPTION}

\item{to}{PARAM_DESCRIPTION}

\item{collection}{PARAM_DESCRIPTION}

\item{url}{PARAM_DESCRIPTION}

\item{client}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION}

\item{as_data_frame}{PARAM_DESCRIPTION, Default: TRUE}

\item{with_geometry}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_transform}}, \code{\link[sf]{st_geometry}}, \code{\link[sf]{st_bbox}}, \code{\link[sf]{st}}, \code{\link[sf]{st_as_sfc}}, \code{\link[sf]{geos_binary_pred}}, \code{\link[sf]{geos_measures}}, \code{\link[sf]{geos_binary_ops}}
 \code{\link[geojsonsf]{sfc_geojson}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr2]{request}}, \code{\link[httr2]{req_body}}, \code{\link[httr2]{req_auth_bearer_token}}, \code{\link[httr2]{req_oauth_client_credentials}}, \code{\link[httr2]{req_perform}}, \code{\link[httr2]{resp_body_raw}}
 \code{\link[lutz]{tz_lookup_coords}}
 \code{\link[lubridate]{with_tz}}
}
