% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueCatalog.R
\name{UniqueCatalog}
\alias{UniqueCatalog}
\title{Produce image catalog without multiple entries per date}
\usage{
UniqueCatalog(
  imageCatalog,
  by = c("areaCoverage", "tileCloudCover"),
  keep = names(imageCatalog)
)
}
\arguments{
\item{imageCatalog}{\code{data.frame} as returned by the \code{SearchCatalog} function.}

\item{by}{character indicating which attribute is used to select the best image per date.
Can be either "areaCoverage" or "tileCloudCover".}

\item{keep}{list of columns to keep in output. Default: all columns in input.}
}
\value{
\code{data.frame} with one row per date.
}
\description{
Sometimes several images could be available for the given day. It can be useful to have a list
    where for any given day there is just one row in the list. This unique row can be selected to represent either
    the least cloud coverage or the biggest coverage of the are of interest.
}
\details{
The returned \code{data.frame} has the same columns as the input catalog.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[CDSE]{SearchCatalog}}
}
